+++ src/lib/Containers/__tests__/index-tests.tsx
@@ -21,7 +21,6 @@ Array [
   "Inquiry",
   "MyProfile",
   "RegistrationFlow",
-  "Sale",
   "Show",
   "WorksForYou",
 ]